Vectorized documents and related vocabulary
Added two files: one containing a selection of the documents in the Reuters corpus as a vectors. The other the corresponding vocabulary to retrieving the words from the vector elements if that is useful. This selection of documents doesn't seem to be exactly the same as the authors did in the LDA paper so I'll maybe change that later on. But these files are correct so you can start working on the VI with them. 

PS. I'll add the code that generated them tomorrow :)